refactor system settings subpages
