+++ apps/web/src/i18n/locales/en.json
@@ -563,6 +563,21 @@
     "save": "Save",
     "saved": "Saved",
     "saveError": "Save failed",
+    "settingsNavOverview": "Overview",
+    "settingsNavOperations": "Operations",
+    "settingsNavRuntime": "AI and agents",
+    "settingsNavConfigSources": "Config sources",
+    "settingsNavKeywords": "Keywords",
+    "settingsNavLocations": "Locations",
+    "settingsOverviewLabel": "Settings overview",
+    "settingsOverviewTitle": "Open each system settings area in its own page.",
+    "settingsOverviewPageDescription": "Use the local settings navigation above to move directly to runtime, config, keyword, or location management without scrolling through one giant page.",
+    "operationsPageDescription": "Live diagnostics and manual collection controls.",
+    "runtimePageDescription": "Inspect AI connectivity and the screening pipeline.",
+    "runtimePageMeta": "{{count}} bonded agents",
+    "configSourcesPageDescription": "Inspect read-only prompt and configuration sources.",
+    "keywordsPageDescription": "Manage editable keywords and review derived brand data.",
+    "locationsPageDescription": "Control which system location chips are visible in the UI.",
     "customKeywords": "Custom Keywords",
     "customKeywordsDescription": "Manage custom tags available in resume keyword chips",
     "addCustomKeyword": "Add Keyword",
@@ -587,6 +602,19 @@
     "configSourceResolvedLocale": "Resolved Source Locale",
     "configSourceFallbackEnabled": "Fallback to zh-Hans",
     "configSourceFallbackDisabled": "No fallback",
+    "systemLocationConfigTitle": "System location config",
+    "systemLocationConfigDescription": "Backed by Job5156 location data with per-chip visibility controls.",
+    "systemLocationSearchPlaceholder": "Search locations by name, parent, or level",
+    "systemLocationKeyword": "Location",
+    "systemLocationLevel": "Level",
+    "systemLocationParent": "Parent",
+    "systemLocationVisible": "Status",
+    "systemLocationVisibleShown": "Visible",
+    "systemLocationVisibleHidden": "Hidden",
+    "systemLocationProvince": "Province",
+    "systemLocationCity": "City",
+    "systemLocationShowAction": "Show",
+    "systemLocationHideAction": "Hide",
     "readOnly": "Read-only",
     "brandKeywords": "Brand Keywords",
     "brandKeywordsDescription": "Equipment brand data (read-only, source: brands.json)",
