+++ apps/web/src/i18n/locales/zh-Hans.json
@@ -563,6 +563,21 @@
     "save": "保存",
     "saved": "已保存",
     "saveError": "保存失败",
+    "settingsNavOverview": "概览",
+    "settingsNavOperations": "运行操作",
+    "settingsNavRuntime": "AI 与 Agent",
+    "settingsNavConfigSources": "配置来源",
+    "settingsNavKeywords": "关键词",
+    "settingsNavLocations": "地区配置",
+    "settingsOverviewLabel": "设置概览",
+    "settingsOverviewTitle": "将系统设置拆分到独立页面中查看。",
+    "settingsOverviewPageDescription": "使用上方的本地设置导航，直接进入运行状态、配置来源、关键词或地区配置页面，不再需要在一个超长页面中来回滚动。",
+    "operationsPageDescription": "查看实时诊断信息，并手动触发采集操作。",
+    "runtimePageDescription": "查看 AI 连通状态与筛选 Agent 流程。",
+    "runtimePageMeta": "已绑定环境的 Agent：{{count}} 个",
+    "configSourcesPageDescription": "查看只读的提示词和配置来源。",
+    "keywordsPageDescription": "管理可编辑关键词，并查看派生出来的品牌数据。",
+    "locationsPageDescription": "控制系统地区标签在 UI 中的显示与隐藏。",
     "customKeywords": "自定义关键词",
     "customKeywordsDescription": "管理可在简历页面关键词芯片中使用的自定义标签",
     "addCustomKeyword": "添加关键词",
@@ -587,6 +602,19 @@
     "configSourceResolvedLocale": "实际来源语言",
     "configSourceFallbackEnabled": "已回退到 zh-Hans",
     "configSourceFallbackDisabled": "未发生回退",
+    "systemLocationConfigTitle": "系统地区配置",
+    "systemLocationConfigDescription": "来源于 Job5156 地区数据，可配置展开标签显示或隐藏。",
+    "systemLocationSearchPlaceholder": "搜索地区（名称 / 上级 / 层级）",
+    "systemLocationKeyword": "地区",
+    "systemLocationLevel": "层级",
+    "systemLocationParent": "上级",
+    "systemLocationVisible": "状态",
+    "systemLocationVisibleShown": "显示",
+    "systemLocationVisibleHidden": "隐藏",
+    "systemLocationProvince": "省级",
+    "systemLocationCity": "城市",
+    "systemLocationShowAction": "显示",
+    "systemLocationHideAction": "隐藏",
     "readOnly": "只读",
     "brandKeywords": "品牌关键词",
     "brandKeywordsDescription": "设备品牌数据（只读，来源：brands.json）",
