+++ apps/web/src/i18n/locales/zh-Hant.json
@@ -563,6 +563,21 @@
     "save": "儲存",
     "saved": "已儲存",
     "saveError": "儲存失敗",
+    "settingsNavOverview": "概覽",
+    "settingsNavOperations": "執行操作",
+    "settingsNavRuntime": "AI 與 Agent",
+    "settingsNavConfigSources": "配置來源",
+    "settingsNavKeywords": "關鍵詞",
+    "settingsNavLocations": "地區配置",
+    "settingsOverviewLabel": "設定概覽",
+    "settingsOverviewTitle": "將系統設定拆分到獨立頁面中查看。",
+    "settingsOverviewPageDescription": "使用上方的本地設定導覽，直接進入執行狀態、配置來源、關鍵詞或地區配置頁面，不必再在超長頁面中來回捲動。",
+    "operationsPageDescription": "查看即時診斷資訊，並手動觸發採集操作。",
+    "runtimePageDescription": "查看 AI 連線狀態與篩選 Agent 流程。",
+    "runtimePageMeta": "已綁定環境的 Agent：{{count}} 個",
+    "configSourcesPageDescription": "查看唯讀的提示詞與配置來源。",
+    "keywordsPageDescription": "管理可編輯關鍵詞，並查看衍生出的品牌資料。",
+    "locationsPageDescription": "控制系統地區標籤在 UI 中的顯示與隱藏。",
     "customKeywords": "自定義關鍵詞",
     "customKeywordsDescription": "管理可在簡歷頁面關鍵詞晶片中使用的自定義標籤",
     "addCustomKeyword": "新增關鍵詞",
@@ -587,6 +602,19 @@
     "configSourceResolvedLocale": "實際來源語言",
     "configSourceFallbackEnabled": "已回退到 zh-Hans",
     "configSourceFallbackDisabled": "未發生回退",
+    "systemLocationConfigTitle": "系統地區配置",
+    "systemLocationConfigDescription": "來源於 Job5156 地區資料，可配置展開標籤顯示或隱藏。",
+    "systemLocationSearchPlaceholder": "搜尋地區（名稱 / 上級 / 層級）",
+    "systemLocationKeyword": "地區",
+    "systemLocationLevel": "層級",
+    "systemLocationParent": "上級",
+    "systemLocationVisible": "狀態",
+    "systemLocationVisibleShown": "顯示",
+    "systemLocationVisibleHidden": "隱藏",
+    "systemLocationProvince": "省級",
+    "systemLocationCity": "城市",
+    "systemLocationShowAction": "顯示",
+    "systemLocationHideAction": "隱藏",
     "readOnly": "唯讀",
     "brandKeywords": "品牌關鍵詞",
     "brandKeywordsDescription": "設備品牌資料（唯讀，來源：brands.json）",
